fix(main-page.tsx): Fix data not being placed correctly when switching to record using different interface
