+++ package.json
@@ -22,7 +22,8 @@
     "react-dom": "^19.1.0",
     "react-hook-form": "^7.62.0",
     "react-icons": "^5.5.0",
-    "react-resizable-panels": "^3.0.5"
+    "react-resizable-panels": "^3.0.5",
+    "zustand": "^5.0.8"
   },
   "devDependencies": {
     "@tauri-apps/cli": "^2",
