+++ src/common/types/api-interface-types.ts
@@ -1,0 +1,4 @@
+export enum InterfaceType {
+  HTTP = 'HTTP',
+  WS = 'WS',
+}
