+++ src/pages/main-page/components/http-display/components/http-request-form/http-request-form.tsx
@@ -6,8 +6,8 @@ import HeadersControl from './components/headers-control';
 import BodyControl from './components/body-control';
 import UrlControl from './components/url-control';
 import { HiArchiveBoxArrowDown } from 'react-icons/hi2';
-import { InterfaceType } from '../../../../types';
-import InterfaceTypeControl from '../../../interface-type-control/interface-type-control';
+import InterfaceTypeSelect from '../../../interface-type-select/interface-type-select';
+import { InterfaceType } from '../../../../../../common/types/api-interface-types';
 
 type HTTPRequestFormProps = {
   interfaceForm: UseFormReturn<HTTPInputs>;
@@ -49,7 +49,7 @@ const HTTPRequestForm = ({
       }}
     >
       <Box display="flex" gap={1}>
-        <InterfaceTypeControl control={control} />
+        <InterfaceTypeSelect />
         <Box flexGrow={1}>
           <UrlControl control={control} />
         </Box>
